updated Data_Ingestion module
